+++ lib/util.lua
@@ -1,9 +1,0 @@
-local M = {}
-
-function M.find_attr(attrs, name)
-  for _, a in ipairs(attrs or {}) do
-    if a.name == name then return a end
-  end
-end
-
-return M